Calculate distance along the path from any position
We need to know the distance along the path so that towers can
target the enemy closest or farthest from reaching the exit.
In a tower defense game with fixed paths, this is easy. Just get
the progress along the path. In mazing games with pathfinding,
this is also easy. It falls out of pathfinding. For us, its
mildly tricky.

Enemies in this game move along lanes as if they were track
athletes, except lanes are continuous instead of discrete. This
presents a slight wrinkle at corners, where corner tiles exhibit
two different rotation amounts. Dealing with this in a simple
way is a bit hacky, but the overall system is nice and compart-
ementalized.