Move sprite rendering to js-calls-rust instead of rust-calls-js
It's hard to tell, and likely premature anyway, but this change
seems to have increasedd performance, since it was a good
opportunity to use a texture atlas instead of separate textures.
It also seemed like a good chance to use ParticleContainer, but
that bugged out for unknown reasons. But a regular Container should
be just fine, maybe even preferable, for the small number of sprites
that I have.

Also added a visual missile tower like Onslaught 2. And began
playing around with fragment shaders for showing range and for
animating ice.